Move jdbc configurations
